% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theorical_comparison.R
\name{.add_EQ5D_utilities}
\alias{.add_EQ5D_utilities}
\title{.add_EQ5D_utilities}
\usage{
.add_EQ5D_utilities(
  df,
  value_sets,
  version,
  dim.names = c("mo", "sc", "ua", "pd", "ad"),
  colnames = NULL
)
}
\arguments{
\item{df}{A data frame to which the utilities will be added.}

\item{value_sets}{A character vector specifying the country value sets for the given EQ5D version.}

\item{version}{A character string specifying the EQ5D version. Valid versions are "5L", "3L", "XW", and "XWR".}

\item{dim.names}{A vector of dimension names to identify dimension columns in the df data frame}

\item{colnames}{(Optional) A character vector specifying the column names for the added utilities.}
}
\value{
The df data frame with the added utility columns.
}
\description{
This function calculates utilities for the provided EQ5D version and value sets.
}
