% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.calculate_weighted_statistics}
\alias{.calculate_weighted_statistics}
\title{.calculate_weighted_statistics}
\usage{
.calculate_weighted_statistics(
  data_array,
  data_margin = 2,
  quantile_levels = c(min = 0, `2.5\%` = 0.025, `25\%` = 0.25, median = 0.5, `75\%` =
    0.75, `97.5\%` = 0.975, max = 1)
)
}
\arguments{
\item{data_array}{A data array where calculations will be performed on each column.}

\item{quantile_levels}{A named vector of probabilities for which quantiles are required.}
}
\value{
A data frame with columns for each calculated statistic, including mean, standard deviation, upper and lower bounds, and user-specified quantiles.
}
\description{
This function calculate the mean, standard deviation, and quantiles for each column in a provided data array.
}
\keyword{internal}
