% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.compute_ratios}
\alias{.compute_ratios}
\title{.compute_ratios}
\usage{
.compute_ratios(df, combinations, utility_columns)
}
\arguments{
\item{df}{A data frame containing the utility variables for which ratios are to be computed.}

\item{combinations}{A matrix variable specifying the utility_columns combinations.}

\item{utility_columns}{A character vector specifying the names of the utility variables in the data frame.}
}
\value{
A data frame with additional columns corresponding to the computed ratios.
}
\description{
The function takes in a data frame and a set of utility variables. For each pair of utility variables, 
    it computes their ratio and adds a new column to the data frame. The new columns are named in the format "F_ratio_Var1_Var2".
}
\keyword{internal}
