% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.cut_variable}
\alias{.cut_variable}
\title{.cut_variable}
\usage{
.cut_variable(variable, breaks)
}
\arguments{
\item{variable}{A numeric vector that you want to cut into intervals.}

\item{breaks}{A numeric vector specifying the breakpoints for cutting the variable.}
}
\value{
A factor vector representing the intervals into which the variable has been cut.
}
\description{
This function cuts a numeric variable into intervals based on the provided breaks.
}
