% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.f_stat_from_df}
\alias{.f_stat_from_df}
\title{.f_stat_from_df}
\usage{
.f_stat_from_df(df, include_p = FALSE)
}
\arguments{
\item{df}{A data frame containing one or more numeric columns and a reference column named 'gr'.}

\item{include_p}{A logical value indicating whether to include the p-values in  the result.}
}
\value{
If include_p is FALSE, a named numeric vector of F-statistics for each column against the 'gr' column. 
  If include_p is TRUE, a matrix with  two rows containing F-statistics and p-values, respectively.
}
\description{
The function is designed to compute F-statistic(s) using one-way ANOVA for each numeric column in the provided data frame against a reference column named 'gr'.
}
\keyword{internal}
