% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.factorize_variable}
\alias{.factorize_variable}
\title{.factorize_variable}
\usage{
.factorize_variable(df, weight_column, variant_fun, breaks = NULL)
}
\arguments{
\item{df}{A data frame containing the variable to be factorized.}

\item{weight_column}{A string specifying the name of the column in the data frame that contains the variable to be factorized.}

\item{variant_fun}{A function that will be applied to the variable for factorization.}

\item{breaks}{An optional numeric vector specifying the breakpoints for cutting the variable, if applicable.}
}
\value{
A factor or numeric vector representing the factorized variable.
}
\description{
This function factorizes a numeric variable in a data frame based on a specified variant function and optional breaks.
}
\keyword{internal}
