% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.flatten_group_to_df}
\alias{.flatten_group_to_df}
\title{.flatten_group_to_df}
\usage{
.flatten_group_to_df(x)
}
\arguments{
\item{x}{A matrix containing the data to be flattened. The matrix should have an attribute "gr" that contains the size for each group.}
}
\value{
A data frame containing the flattened data with an additional 'gr' column indicating the group each row belongs to.
}
\description{
This function takes a matrix with an attribute "gr" representing group sizes and flattens it into a data frame, adding a 'gr' column to indicate the group each row belongs to.
}
\keyword{internal}
