% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theorical_comparison.R
\name{.get_EQ5D_value_sets}
\alias{.get_EQ5D_value_sets}
\title{.get_EQ5D_value_sets}
\usage{
.get_EQ5D_value_sets(version, value_sets, value_sets_XW)
}
\arguments{
\item{version}{A character string specifying the EQ5D version.}

\item{value_sets}{A character vector specifying the country value sets for the given EQ5D version.}

\item{value_sets_XW}{A character vector specifying the country value sets for the crosswalk EQ5D version.}
}
\value{
A list containing:
\itemize{
  \item \code{utilityColumn}: A character vector with the names of the added utility columns.
  \item \code{df}: A data frame containing all generated EQ states with their utilities.
}
}
\description{
This function generates all EQ states based on the provided EQ5D version and adds utilities.
}
\keyword{internal}
