% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theorical_comparison.R
\name{.get_EQ5D_version}
\alias{.get_EQ5D_version}
\title{.get_EQ5D_version}
\usage{
.get_EQ5D_version(df, stateColumn)
}
\arguments{
\item{df}{A data frame containing EQ5D health states.}

\item{stateColumn}{A character string specifying the name of the column in the df data frame that contains the EQ5D health states.}
}
\value{
A character string indicating the version of EQ5D used. Valid return values are "3L" for EQ5D-3L, "5L" for EQ5D-5L, or NULL if the version cannot be determined.
}
\description{
This function determines the version of EQ5D used based on the provided stateColumn of a data frame.
}
\keyword{internal}
