% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.get_Fstatistics_interpretation}
\alias{.get_Fstatistics_interpretation}
\title{.get_Fstatistics_interpretation}
\usage{
.get_Fstatistics_interpretation(errorbar_plot, utility_combinations = NULL)
}
\arguments{
\item{errorbar_plot}{A ggplot object containing an error bar plot.}

\item{utility_combinations}{A matrix or data frame of utility combinations corresponding to the x-axis labels (default is NULL).}
}
\value{
A list containing the interpretation results for each combination of types in the F-statistic plot.
        Each list element contains details about significance, mean, confidence interval.
}
\description{
This function extracts and interprets the results from a given F-statistics error bar plot.
}
\keyword{internal}
