% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.makeWeights}
\alias{.makeWeights}
\title{.makeWeights}
\usage{
.makeWeights(
  x,
  pointval = 100,
  rng = c(0:100),
  variant_fun = .makeWeightsMixed
)
}
\arguments{
\item{x}{A numeric vector of input values.}

\item{pointval}{A numeric point value within the range of x, default is 100.}

\item{rng}{A numeric vector indicating the range of x, default is 0 to 100.}

\item{variant_fun}{A function to calculate weights. It can be makeWeightsTriangular, makeWeightsGradient, makeWeightsMixed, or a custom function name. Default is makeWeightsTriangular.}
}
\value{
A numeric vector of calculated and scaled weights.
}
\description{
A function to calculate weights with specified variant function.
}
\keyword{internal}
