% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.makeWeightsGradient}
\alias{.makeWeightsGradient}
\title{.makeWeightsGradient}
\usage{
.makeWeightsGradient(x, pointval, rng)
}
\arguments{
\item{x}{A numeric vector of input values.}

\item{pointval}{A numeric point value within the range of x.}

\item{rng}{A numeric vector indicating the range of x.}
}
\value{
A numeric vector of calculated weights.
}
\description{
A function to calculate weights based on a gradient approach.
}
