% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.merge_adjacent_intervals}
\alias{.merge_adjacent_intervals}
\title{.merge_adjacent_intervals}
\usage{
.merge_adjacent_intervals(intervals)
}
\arguments{
\item{intervals}{A character vector of intervals to be merged.}
}
\value{
A character vector of merged intervals.
}
\description{
This function takes a character vector of intervals (e.g., "0-25", "25-50") and merges adjacent intervals.
}
\keyword{internal}
