% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.plot_F_statististics}
\alias{.plot_F_statististics}
\title{.plot_F_statistics}
\usage{
.plot_F_statististics(
  df,
  utility_columns,
  utility_combinations = NULL,
  graph_title = "",
  x_axis_title = "",
  y_axis_title = "",
  y_min_value = NULL,
  y_max_value = NULL
)
}
\arguments{
\item{df}{A data frame containing the F-statistics.}

\item{utility_columns}{A character vector specifying the names of the utility columns in the data frame.}

\item{graph_title}{A string specifying the title of the graph. Default is an empty string.}

\item{x_axis_title}{A string specifying the title for the x-axis. Default is an empty string.}

\item{y_axis_title}{A string specifying the title for the y-axis. Default is an empty string.}

\item{y_min_value}{A numeric value specifying the minimum limit for the y-axis. Default is NULL.}

\item{y_max_value}{A numeric value specifying the maximum limit for the y-axis. Default is NULL.}
}
\value{
A ggplot object representing the bar plot with error bars.
}
\description{
This function creates a bar plot for F-statistics along with error bars for specified utility columns.
}
\keyword{internal}
