% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.write_Fstatistics_interpretation}
\alias{.write_Fstatistics_interpretation}
\title{.write_Fstatistics_interpretation}
\usage{
.write_Fstatistics_interpretation(errorbar_plot, utility_combinations = NULL)
}
\arguments{
\item{errorbar_plot}{A ggplot object containing an error bar plot.}

\item{utility_combinations}{A matrix or data frame of utility combinations corresponding to the x-axis labels (default is NULL).}
}
\value{
A character vector containing the interpretation paragraphs for each combination of types in the F-statistic plot.
}
\description{
This function generates the interpretation of the results obtained from a F-statistics plot.
}
\keyword{internal}
