% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.write_severity_interpretation}
\alias{.write_severity_interpretation}
\title{.write_severity_interpretation}
\usage{
.write_severity_interpretation(
  ribbon_plot,
  quartiles = c(0, 0.25, 0.5, 0.75, 1),
  elevation_threshold = 0.05,
  slope_threshold = 0.02
)
}
\arguments{
\item{ribbon_plot}{A severity ribbon plot to be interpreted.}

\item{quartiles}{A numeric vector specifying the quartiles for interpretation. Default is c(0, 0.25, 0.5, 0.75, 1).}

\item{elevation_threshold}{A numeric value specifying the threshold for elevation differences. Default is 0.05.}

\item{slope_threshold}{A numeric value specifying the threshold for slope differences. Default is 0.02.}
}
\value{
A character vector containing the interpretation paragraphs for each combination of types in the ribbon plot.
}
\description{
This function generates the interpretation of the results obtained from a severity ribbon plot.
}
\keyword{internal}
