% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.get_Fstatistics_interpretation}
\alias{.get_Fstatistics_interpretation}
\title{.get_Fstatistics_interpretation}
\usage{
.get_Fstatistics_interpretation(errorbar_plot)
}
\arguments{
\item{errorbar_plot}{A numeric vector specifying the quartiles for interpretation. Default is c(0, 0.25, 0.5, 0.75, 1).}

\item{ribbon_plot}{A F-statistics error bar plot to be interpreted.}
}
\value{
A list containing the interpretation results for each combination of types in the F-statistic plot.
        Each list element contains details about sinficance, mean, confidence interval.
}
\description{
This function extracts and interprets the results from a given F-statistics error bar plot.
}
\keyword{internal}
