% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.write_Fstatistics_interpretation}
\alias{.write_Fstatistics_interpretation}
\title{.write_Fstatistics_interpretation}
\usage{
.write_Fstatistics_interpretation(errorbar_plot, utility_combinations = NULL)
}
\arguments{
\item{intepretation_results}{A list containing the interpretation results for the F-statistics plot.}
}
\value{
A character vector containing the interpretation paragraphs for each combination of types in the ribbon plot.
}
\description{
This function generates the interpretation of the results obtained from a F-statistics plot.
}
