% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_comparison.R
\name{.write_severity_intepretation}
\alias{.write_severity_intepretation}
\title{.write_severity_intepretation}
\usage{
.write_severity_intepretation(intepretation_results)
}
\arguments{
\item{intepretation_results}{A list containing the interpretation results for each combination of types in the ribbon plot.}
}
\value{
A character vector containing the interpretation paragraphs for each combination of types in the ribbon plot.
}
\description{
This function generates the interpretation of the results obtained from a severity ribbon plot.
}
\keyword{internal}
